Increase the size of the playout variable.

+++ Mode3TX.h
@@ -41,7 +41,7 @@ public:
 
 private:
   CRingBuffer<uint8_t>             m_fifo;
-  uint8_t                          m_playOut;
+  uint16_t                         m_playOut;
   arm_fir_interpolate_instance_q15 m_modFilter;
   q15_t                            m_modState[16U];    // blockSize + phaseLength - 1, 4 + 9 - 1 plus some spare
   CIL2PTX                          m_frame;
